Rewind option now scrolls to track edge or beginning before going over
Changes `Controller` component behavior, so affects both arrows and
programmatic navigation.

This fixes an issue with sliders where `perMove` is
above 1 and the number of slides isn't divisible by `perMove`. In these
cases, rewind would skip showing slides as follows:
- In case of one whole page and one partial page, next arrow button did
nothing. Previous arrow button rewound to the edge correctly on the
first press but afterwards would not go back to start. In this case the next
arrow button would go the start.
- In case of multiple pages where the last one is partial, the last page
would be skipped when navigating using the next arrow button, and the first
page would be skipped when navigating using the previous arrow button.

The fix causes rewind to first go the edge if not yet at the edge for
positive indices, and to first go to the beginning if not yet at the
beginning for negative indices.

Note that this also changes the behavior of `rewind` for sliders where
`perMove` is 1 as indices above the edge or below zero will now go
to the edge or zero unless already there. I feel like this is more in
line with user expectations as it doesn't inadvertently skip slides,
but the change should probably be considered breaking.

If the change is too drastic, I suppose the same could be achieved
by changing the behavior of the `Arrows` component to take this into
account.